Декомпозирует компонент Main-Page, чтобы не перерисовывались header и навигация
