+++ project/src/components/empty-main-page/empty-main-page.tsx
@@ -1,15 +1,20 @@
-type EmptyMainPageProps = {
-  cityName: string;
-};
+import { useAppSelector } from '../../hooks/';
 
-function EmptyMainPage({cityName}: EmptyMainPageProps): JSX.Element {
+function EmptyMainPage(): JSX.Element {
+  const cityName = useAppSelector((state) => state.city);
+
   return (
-    <section className="cities__no-places">
-      <div className="cities__status-wrapper tabs__content">
-        <b className="cities__status">No places to stay available</b>
-        <p className="cities__status-description">We could not find any property available at the moment in {cityName}</p>
+    <div className="cities">
+      <div className="cities__places-container cities__places-container--empty container">
+        <section className="cities__no-places">
+          <div className="cities__status-wrapper tabs__content">
+            <b className="cities__status">No places to stay available</b>
+            <p className="cities__status-description">We could not find any property available at the moment in {cityName}</p>
+          </div>
+        </section>
+        <div className="cities__right-section"></div>
       </div>
-    </section>
+    </div>
   );
 }
